requirements.txt
alabaster==0.7.12
amqp==5.0.7
anyio==3.4.0
argon2-cffi==20.1.0
arrow==1.2.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.9.1
backcall==0.2.0
baytune==0.4.0
billiard==3.6.4.0
binaryornot==0.4.4
bleach==3.2.1
celery==5.2.1
certifi==2020.6.20
cffi==1.14.2
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cookiecutter==1.7.3
copulas==0.3.3
coverage==6.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.17.1
entrypoints==0.3
h5py==2.10.0
idna==3.3
imagesize==1.3.0
imblearn-0.0
imbalanced-learn==0.8.1
imutils==0.5.3
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.16.0
json5==0.9.6
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.1
jupyterlab-server==2.9.0
jupyterlab-widgets==1.0.2
Keras==2.4.3
kiwisolver==1.2.0
kombu==5.2.2
make==0.1.6.post2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mrcnn==0.2
nbclassic==0.3.4
nbclient==0.5.0
nbconvert==6.0.3
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.6.3
notebook==6.1.4
numpy==1.21.4
opencv-python==4.4.0.42
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==1.0.0
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2-binary==2.9.2
py==1.11.0
pycparser==2.20
pydot==1.4.2
Pygments==2.7.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.1
python-logstash==0.4.6
python-slugify==5.0.2
pytz==2020.1
#pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.2.2
QtPy==1.11.3
redis==3.5.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.5.2
seaborn==0.11.2
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==4.3.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
terminado==0.9.0
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tomli==2.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.4
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.5.2
xgboost==1.5.1
